// postnet-client/packages/postnet-api/src/model/post-vo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * POSTNET API
 * API k semestrální práce z PIA
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Přepravka pro příspěvek
 * @export
 * @interface PostVO
 */
export interface PostVO {
    /**
     * 
     * @type {string}
     * @memberof PostVO
     */
    author?: string;
    /**
     * 
     * @type {string}
     * @memberof PostVO
     */
    creationDateTime?: string;
    /**
     * 
     * @type {string}
     * @memberof PostVO
     */
    content?: string;
}