// postnet-client/packages/postnet-api/src/model/user-vo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * POSTNET API
 * API k semestrální práce z PIA
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Přepravka pro údaje o uživateli
 * @export
 * @interface UserVO
 */
export interface UserVO {
    /**
     * 
     * @type {number}
     * @memberof UserVO
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof UserVO
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof UserVO
     */
    name?: string;
    /**
     * 
     * @type {boolean}
     * @memberof UserVO
     */
    admin?: boolean;
}